7. Using the session instead of the 'Opt' model in the 'register' and 'check_opt' view and unregister it.

+++ account/views.py
@@ -5,7 +5,8 @@ from .models import User, Otp
 from kavenegar import KavenegarAPI
 from random import randint
 from django.urls import reverse
-
+from datetime import timedelta
+from django.utils import timezone
 
 SMS = KavenegarAPI(apikey='484236523838636B4178655269387331566A7932673638786D6C6155376B7944554137435A3973424335733D')
 
@@ -39,10 +40,16 @@ def user_register(request):
         if form.is_valid():
             cd = form.cleaned_data
             rand_code = randint(1000, 9999)
+            print(rand_code)
             '''response = SMS.sms_send(
                 {'receptor': str(cd.get('phone')), 'message': f'Multi Shop\nVerification code: {rand_code}'})'''
-            otp = Otp.objects.create(phone=cd.get('phone'), rand_code=rand_code, full_name=cd['full_name'], password=cd['password'])
-            return redirect(reverse('account:check_otp') + f'?token={otp.token}')
+            # otp = Otp.objects.create(phone=cd.get('phone'), rand_code=rand_code, full_name=cd['full_name'], password=cd['password'])
+            request.session.set_expiry(120)
+            request.session['opt'] = {'phone': cd.get('phone'),
+                                      'code': rand_code,
+                                      'full_name': cd.get('full_name'),
+                                      'password': cd.get('password')}
+            return redirect(reverse('account:check_otp'))  # + f'?token={otp.token}')
     else:
         form = UserRegistrationForm()
 
@@ -55,21 +62,37 @@ def check_opt(request):
 
     if request.method == 'POST':
         form = OtpForm(request.POST)
-        token = request.GET.get('token')
-        if form.is_valid():
-            try:
-                user_info = Otp.objects.get(token=token, rand_code=form.cleaned_data['code'])
-                user = User.objects.create_user(phone=user_info.phone, full_name=user_info.full_name, password=user_info.password)
-                user_info.delete()
-                login(request, user)
-                return redirect('/')
-            except:
-                form.add_error('code', 'Verification code is invalid')
+        user_info = request.session.get('opt', {})
+        if form.is_valid() and user_info and int(form.cleaned_data.get('code')) == user_info.get('code'):
+            user = User.objects.create_user(phone=user_info['phone'], full_name=user_info['full_name'],
+                                            password=user_info['password'])
+            login(request, user)
+            return redirect('/')
+        else:
+            form.add_error('code', 'Verification code is invalid')
     else:
         form = OtpForm()
 
     return render(request, 'account/check_otp.html', {'form': form})
 
 
-
-
+'''
+if request.method == 'POST':
+        form = OtpForm(request.POST)
+        token = request.GET.get('token')
+        if form.is_valid():
+            try:
+                user_info = Otp.objects.get(token=token, rand_code=form.cleaned_data['code'])
+                if user_info.expiration_date + timedelta(minutes=2) < timezone.now():
+                    user_info.delete()
+                    del user_info
+                user = User.objects.create_user(phone=user_info.phone, full_name=user_info.full_name, password=user_info.password)
+                user_info.delete()
+                login(request, user)
+                return redirect('/')
+            except:
+                otp = Otp.objects.filter(token=token)
+                if otp.exists():
+                    otp.delete()
+                form.add_error('code', 'Verification code is invalid')
+'''
